pyproject.toml
[project]
name = "figh_skspark"
version = "0.1.0"
description = "version controlled configuration storage"
authors = [
    {name = "skspark", email = "wondershine.one@gmail.com"},
]
dependencies = [
    "typing>=3.7.4.3",
    "PyGithub>=1.59.0",
    "PyYaml>=6.0.1",
    "pydantic>=2.0.3",
]
requires-python = ">=3.7"
readme = "README.md"
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm.dev-dependencies]
dev = [
    "pytest>=7.4.0",
    "wheel>=0.40.0",
    "twine>=4.0.2",
]
